requirements.txt
# 第一阶段图像处理依赖包
# 基础科学计算
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0

# 图像处理
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
Pillow>=8.3.0
scikit-image>=0.18.0

# 机器学习（预留）
scikit-learn>=1.0.0

# 数据处理
pandas>=1.3.0

# 配置文件处理
PyYAML>=5.4.0

# 进度条和日志
tqdm>=4.62.0
loguru>=0.5.0

# 测试框架
pytest>=6.2.0
pytest-cov>=2.12.0

# Jupyter支持
jupyter>=1.0.0
ipykernel>=6.0.0

# 可视化增强
seaborn>=0.11.0
plotly>=5.0.0

# 3D模型处理
trimesh>=3.15.0
open3d>=0.15.0
vtk>=9.1.0

# 开发工具
black>=21.0.0
flake8>=3.9.0